+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-# Apr 30 2021
+### May 1, 2021 
+
+- you can now insert blocks above and below your current block via command palette (Type `/` to use it) or by pressing `Command-Enter` (below) or `Command-Shift-Enter` (above). 
+
+### Apr 30 2021
 
 - A new inline command palette! Type `/` to use it.